moving debug menu access to only the debug sourceset and providing an injectable interface to decouple from the direct access
