+++ vector/src/debug/AndroidManifest.xml
@@ -9,6 +9,7 @@
         <activity android:name=".features.debug.settings.DebugPrivateSettingsActivity" />
         <activity android:name=".features.debug.sas.DebugSasEmojiActivity" />
         <activity android:name=".features.debug.features.DebugFeaturesSettingsActivity" />
+        <activity android:name=".features.debug.DebugMenuActivity" />
 
         <activity
             android:name="com.facebook.flipper.android.diagnostics.FlipperDiagnosticActivity"
