+++ vector/src/debug/java/im/vector/app/features/debug/di/DebugModule.kt
@@ -16,16 +16,31 @@
 
 package im.vector.app.features.debug.di
 
+import android.content.Context
+import android.content.Intent
 import dagger.Binds
 import dagger.Module
+import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import im.vector.app.core.platform.DebugReceiver
+import im.vector.app.features.debug.DebugMenuActivity
+import im.vector.app.features.navigation.DebugNavigator
 import im.vector.app.receivers.VectorDebugReceiver
 
 @InstallIn(SingletonComponent::class)
 @Module
 abstract class DebugModule {
+
+    companion object {
+
+        @Provides
+        fun providesDebugNavigator() = object : DebugNavigator {
+            override fun openDebugMenu(context: Context) {
+                context.startActivity(Intent(context, DebugMenuActivity::class.java))
+            }
+        }
+    }
 
     @Binds
     abstract fun bindsDebugReceiver(receiver: VectorDebugReceiver): DebugReceiver
